backend/requirements.txt
# Django core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
argon2-cffi==23.1.0

# Task queue
celery==5.3.6
redis==5.0.4

# HTTP client for AI API
httpx==0.27.0

# WSGI server
gunicorn==22.0.0

# Utilities
python-decouple==3.8
